:sparkles: create review and many to one relationships

+++ src/main/java/org/inrikys/domain/models/User.java
@@ -2,7 +2,7 @@ package org.inrikys.domain.models;
 
 public class User {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
 
@@ -11,13 +11,13 @@ public class User {
         this.email = email;
     }
 
-    public User(Integer id, String name, String email) {
+    public User(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
